update readme and load_pretrained
