+++ paddlemix/examples/blip2/run_eval_caption.py
@@ -24,10 +24,10 @@ import paddle
 import paddle.distributed as dist
 from paddle.distributed import fleet
 from paddle.distributed.fleet.meta_parallel import get_rng_state_tracker
-from paddlenlp.trainer import PdArgumentParser, TrainingArguments, get_last_checkpoint
 
 from paddlemix.datasets import load_dataset
-from paddlemix.examples.blip2.utils import BlipCollator, create_tokenizer
+from paddlemix.examples.blip2.utils import BlipCollator, create_tokenizer, load_model
+from paddlemix.models.blip2.configuration import Blip2Config
 from paddlemix.models.blip2.modeling import Blip2ForConditionalGeneration
 from paddlemix.processors.blip_processing import (
     Blip2Processor,
@@ -36,6 +36,7 @@ from paddlemix.processors.blip_processing import (
 )
 from paddlemix.trainer.blip2_trainer import BLIP2Trainer as Trainer
 from paddlemix.utils.log import logger
+from paddlenlp.trainer import PdArgumentParser, TrainingArguments, get_last_checkpoint
 
 
 @dataclass
@@ -120,15 +121,18 @@ class PreTrainingArguments(TrainingArguments):
         default=1, metadata={"help": "Set the number of sharding, enable sharding parallel"}
     )
     pipeline_parallel_degree: int = field(default=1, metadata={"help": "Enable pipeline parallel"})
-    model_path: str = field(
+    load_model_path: str = field(
         default=None,
         metadata={"help": "The path to model if you want to load weights from the specified path"},
     )
 
 
-def create_model(config):
-    # blip2_config = Blip2ForConditionalGeneration(onfig.model_name_or_path)
-    model = Blip2ForConditionalGeneration.from_pretrained(pretrained_model_name_or_path=config.model_name_or_path)
+def create_model(config, training_args=None):
+    blip2_config = Blip2Config.from_pretrained(config.model_name_or_path)
+    blip2_config.mp_degree = config.mp_degree
+    blip2_config.gradient_checkpointing = config.gradient_checkpointing
+    model = Blip2ForConditionalGeneration(blip2_config)
+    model.load_pretrained(blip2_config, model, training_args)
     paddle.device.cuda.empty_cache()
     return model
 
@@ -189,6 +193,8 @@ def main():
     model = create_model(model_args)
     logger.info("training_args.use_hybrid_parallel:{}".format(training_args.use_hybrid_parallel))
     # create trainer
+    if training_args.load_model_path is not None:
+        load_model(training_args, model, ckpt_dir=os.path.join(training_args.load_model_path, "model_state.pdparams"))
     trainer = Trainer(
         model=model,
         args=training_args,
