+++ paddlemix/models/blip2/__init__.py
@@ -11,8 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
+from .base_model import *
 from .configuration import *
 from .modeling import *
-
-__all__ = configuration.__all__ + modeling.__all__
