+++ paddlemix/processors/blip_processing.py
@@ -20,6 +20,7 @@ from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
 import PIL
+
 from paddlenlp.transformers.tokenizer_utils_base import (
     BatchEncoding,
     PreTokenizedInput,
@@ -132,7 +133,7 @@ class Blip2Processor(ProcessorMixin):
                 return_tensors=return_tensors,
                 return_token_type_ids=False,
                 max_length=32,
-                padding=True,
+                padding="longest",
                 **kwargs,
             )
             return text_encoding
@@ -154,7 +155,7 @@ class Blip2Processor(ProcessorMixin):
             else:
                 text_encoding = self.text_processor(text, mode=mode)
                 input_encoding = self.tokenizer(
-                    text=text_encoding,
+                    text=text_encoding["input"],
                     return_tensors="pd",
                     padding="longest",
                     truncation=True,
@@ -162,7 +163,7 @@ class Blip2Processor(ProcessorMixin):
                     return_attention_mask=True,
                 )
                 output_encoding = self.tokenizer(
-                    text=text_encoding,
+                    text=text_encoding["output"],
                     return_tensors="pd",
                     padding="longest",
                     truncation=True,
